Updating orientation virtual sensor based on RD-PQ testing
Orientation virtual sensor tested on RD_PQ device based on workaround
on top of the code being submitted here. Workaround involves enabling
accelerometer events for 100ms and then everytime accel event is
received the gyroscope and geomagnetic sensor events are polled and
provided as input to the orientation filter class which computes the
orientation.

Change-Id: Ie30889de3f380422a052d2ba9455586a64ccb640