Adding automatic test file to test all sensors
This file tests all the sensors and for all events of a sensor and prints success or failure based on the output.
Instead of testing each sensor manually, this can be used to test all sensor events automatically.

Patch Set 3: Added checks for the remaining apis. Changed output to a text file. Changed the function name.
Patch Set 5: Added checks for sf_get_data api.

Change-Id: I67318bb61add70d203b5879004c34a9418370c15