Optimization of template structure - Feature merge from devel/tizen branch
This is one of the four patches being submitted to merge the changes made to devel/tizen to tizen  branch.
This contains the feature where the template structure of the vector and matrix classes was changed.

Patch Set 3: Changes related to coding guidelines as per the comments

Build has been checked for armv7l and aarch64 for public repo.

Change-Id: Ia26d66b1ca351a22de92b858f6b60dbfee242d42