Migrating some code to calculate event from proxi_sensor_hal to proxi_sensor
This code basically checks if the value read is for far or near and based on that, assigns a enum value to the event.
This code needs to be moved to proxi_sensor.cpp so that dependency between sensor_internal.h and proxi_sensor_hal.cpp is removed as proxi_sensor_hal.cpp needs to be moved to a separate repo later.

signed-off-by:"Ankur Garg <ankur29.garg@samsung.com>"
Change-Id: I5a088dc8b5de42f445e7c948b9d539e02167e905